requirements.txt
langchain>=0.1.12
langchain-openai==0.3.18
langchain-community==0.3.18
python-dotenv>=1.0.1
openai>=1.12.0
mypy>=1.9.0
pre-commit>=3.6.0 